Add public folder with Flareact favicon inside.
When I run a dev server in Edge Chromium, the console printed an error that said failed to get ./favicon.ico. This kind of error is a sign that the template is incomplete.  So I add public folder along with the favicon.

It also showcase where the public folder should be located in Flareact project.